[FE] fix: 비어있는 guildId 값으로 단일 조회 요청되는 문제 수정 (#69)
